+++ scripts/rollback/rollback-3.3.9.sh
@@ -1,0 +1,86 @@
+#!/bin/bash
+# timestamp: 1653460337092
+
+npm dist-tag add @midwayjs/egg-layer@3.3.6 latest
+npm dist-tag add @midwayjs/egg-layer@3.3.9 beta
+npm dist-tag add @midwayjs/serverless-app@3.3.6 latest
+npm dist-tag add @midwayjs/serverless-app@3.3.9 beta
+npm dist-tag add @midwayjs/bootstrap@3.3.6 latest
+npm dist-tag add @midwayjs/bootstrap@3.3.9 beta
+npm dist-tag add @midwayjs/cross-domain@3.3.6 latest
+npm dist-tag add @midwayjs/cross-domain@3.3.9 beta
+npm dist-tag add @midwayjs/http-proxy@3.3.7 latest
+npm dist-tag add @midwayjs/http-proxy@3.3.9 beta
+npm dist-tag add @midwayjs/i18n@3.3.6 latest
+npm dist-tag add @midwayjs/i18n@3.3.9 beta
+npm dist-tag add @midwayjs/jwt@3.3.6 latest
+npm dist-tag add @midwayjs/jwt@3.3.9 beta
+npm dist-tag add @midwayjs/passport@3.3.6 latest
+npm dist-tag add @midwayjs/passport@3.3.9 beta
+npm dist-tag add @midwayjs/prometheus-socket-io@3.3.6 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@3.3.9 beta
+npm dist-tag add @midwayjs/security@3.3.6 latest
+npm dist-tag add @midwayjs/security@3.3.9 beta
+npm dist-tag add @midwayjs/static-file@3.3.6 latest
+npm dist-tag add @midwayjs/static-file@3.3.9 beta
+npm dist-tag add @midwayjs/task@3.3.6 latest
+npm dist-tag add @midwayjs/task@3.3.9 beta
+npm dist-tag add @midwayjs/upload@3.3.6 latest
+npm dist-tag add @midwayjs/upload@3.3.9 beta
+npm dist-tag add @midwayjs/validate@3.3.6 latest
+npm dist-tag add @midwayjs/validate@3.3.9 beta
+npm dist-tag add @midwayjs/version@3.3.8 latest
+npm dist-tag add @midwayjs/version@3.3.9 beta
+npm dist-tag add @midwayjs/web@3.3.6 latest
+npm dist-tag add @midwayjs/web@3.3.9 beta
+tnpm dist-tag add @midwayjs/egg-layer@3.3.6 latest
+tnpm dist-tag add @midwayjs/egg-layer@3.3.9 beta
+tnpm dist-tag add @midwayjs/serverless-app@3.3.6 latest
+tnpm dist-tag add @midwayjs/serverless-app@3.3.9 beta
+tnpm dist-tag add @midwayjs/bootstrap@3.3.6 latest
+tnpm dist-tag add @midwayjs/bootstrap@3.3.9 beta
+tnpm dist-tag add @midwayjs/cross-domain@3.3.6 latest
+tnpm dist-tag add @midwayjs/cross-domain@3.3.9 beta
+tnpm dist-tag add @midwayjs/http-proxy@3.3.7 latest
+tnpm dist-tag add @midwayjs/http-proxy@3.3.9 beta
+tnpm dist-tag add @midwayjs/i18n@3.3.6 latest
+tnpm dist-tag add @midwayjs/i18n@3.3.9 beta
+tnpm dist-tag add @midwayjs/jwt@3.3.6 latest
+tnpm dist-tag add @midwayjs/jwt@3.3.9 beta
+tnpm dist-tag add @midwayjs/passport@3.3.6 latest
+tnpm dist-tag add @midwayjs/passport@3.3.9 beta
+tnpm dist-tag add @midwayjs/prometheus-socket-io@3.3.6 latest
+tnpm dist-tag add @midwayjs/prometheus-socket-io@3.3.9 beta
+tnpm dist-tag add @midwayjs/security@3.3.6 latest
+tnpm dist-tag add @midwayjs/security@3.3.9 beta
+tnpm dist-tag add @midwayjs/static-file@3.3.6 latest
+tnpm dist-tag add @midwayjs/static-file@3.3.9 beta
+tnpm dist-tag add @midwayjs/task@3.3.6 latest
+tnpm dist-tag add @midwayjs/task@3.3.9 beta
+tnpm dist-tag add @midwayjs/upload@3.3.6 latest
+tnpm dist-tag add @midwayjs/upload@3.3.9 beta
+tnpm dist-tag add @midwayjs/validate@3.3.6 latest
+tnpm dist-tag add @midwayjs/validate@3.3.9 beta
+tnpm dist-tag add @midwayjs/version@3.3.8 latest
+tnpm dist-tag add @midwayjs/version@3.3.9 beta
+tnpm dist-tag add @midwayjs/web@3.3.6 latest
+tnpm dist-tag add @midwayjs/web@3.3.9 beta
+
+# Changes:
+
+#  - @midwayjs/egg-layer: 3.3.6 => 3.3.9
+#  - @midwayjs/serverless-app: 3.3.6 => 3.3.9
+#  - @midwayjs/bootstrap: 3.3.6 => 3.3.9
+#  - @midwayjs/cross-domain: 3.3.6 => 3.3.9
+#  - @midwayjs/http-proxy: 3.3.7 => 3.3.9
+#  - @midwayjs/i18n: 3.3.6 => 3.3.9
+#  - @midwayjs/jwt: 3.3.6 => 3.3.9
+#  - @midwayjs/passport: 3.3.6 => 3.3.9
+#  - @midwayjs/prometheus-socket-io: 3.3.6 => 3.3.9
+#  - @midwayjs/security: 3.3.6 => 3.3.9
+#  - @midwayjs/static-file: 3.3.6 => 3.3.9
+#  - @midwayjs/task: 3.3.6 => 3.3.9
+#  - @midwayjs/upload: 3.3.6 => 3.3.9
+#  - @midwayjs/validate: 3.3.6 => 3.3.9
+#  - @midwayjs/version: 3.3.8 => 3.3.9
+#  - @midwayjs/web: 3.3.6 => 3.3.9